Implement wrench removal on Forge
