+++ forge/src/main/java/owmii/powah/forge/PowahForge.java
@@ -1,6 +1,9 @@
 package owmii.powah.forge;
 
 import dev.architectury.platform.forge.EventBuses;
+import net.minecraftforge.common.MinecraftForge;
+import net.minecraftforge.event.entity.player.PlayerInteractEvent;
+import net.minecraftforge.eventbus.api.Event;
 import net.minecraftforge.fml.ModList;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
@@ -8,6 +11,7 @@ import net.minecraftforge.fml.loading.FMLEnvironment;
 import owmii.powah.Powah;
 import owmii.powah.forge.compat.curios.CuriosCompat;
 import owmii.powah.forge.data.DataEvents;
+import owmii.powah.lib.util.Wrench;
 
 @Mod(Powah.MOD_ID)
 public class PowahForge {
@@ -17,6 +21,16 @@ public class PowahForge {
 
 		Powah.init();
 		modEventBus.addListener(DataEvents::gatherData);
+		MinecraftForge.EVENT_BUS.addListener((PlayerInteractEvent.RightClickBlock event) -> {
+			if (event.getUseBlock() == Event.Result.DENY) {
+				return;
+			}
+			var result = Wrench.removeWithWrench(event.getEntity(), event.getLevel(), event.getHand(), event.getHitVec());
+			if (result != null) {
+				event.setCanceled(true);
+				event.setCancellationResult(result);
+			}
+		});
 
 		if (FMLEnvironment.dist.isClient()) {
 			try {
